compressImageをutilsに移動
