+++ backend/src/services/image/iImageService.ts
@@ -2,5 +2,4 @@ export interface IImageService {
   uploadImage(file: File): Promise<string>;
   getImage(imageUrl: string): Promise<Buffer>;
   isImage(file: File): boolean;
-  compressImage(buffer: Buffer): Promise<Buffer>;
 }
